Maneesh: Calculate surface_normals for each obstacle
Earlier surface_normals was defined for the whole grid. Now surface_normals is a 2D array of shape (N,2) where N is the number of obstacles, and 2 corresponds to X and Y directions (for 2D grid).
Also, since surface_normals require dst_obs to be calculated first, it has been moved to seperate function. dst_obs is calculated once per iteration, and we are passing it to propogation directly.